+++ src/aks-preview/azext_aks_preview/vendored_sdks/azure_mgmt_preview_aks/models.py
@@ -5,5 +5,4 @@
 # license information.
 # --------------------------------------------------------------------------
 from .v2017_07_01.models import *
-from .v2019_04_30.models import *
-from .v2021_08_01.models import *
+from .v2021_09_01.models import *